test(capsule): harden #661 coverage + split host_state tests
Three quick wins from the post-review audit:

1. Accessor precedence unit tests for effective_secret_store and
   effective_capsule_log — pins the accessor contract directly rather
   than relying only on chain tests, so future chain-test refactors
   can't silently break invariant 3.

2. civil_from_days regression tests. The hand-rolled civil-date
   algorithm routes every log file; a bug here misroutes everything.
   Added asserts for epoch (1970-01-01), a leap-day (2000-02-29), a
   month boundary (1970-03-01), a year boundary (1971-01-01), and
   today (2026-04-21). Plus a cross-check that today_date_string's
   formatting matches civil_from_days on the same epoch-seconds value.

3. debug_assert in effective_kv and effective_secret_store: fires in
   debug builds when the invocation principal differs from the capsule
   owner but the corresponding invocation_* field is None. In practice
   unreachable (ScopedKvStore::with_namespace only rejects empty/null-byte
   namespaces, and our format never produces those), but makes the
   degrade-insecure fallback path loud if a future regression breaks
   the invariant. Zero cost in release.

Also: host_state.rs crossed the 1000-line CI threshold after the
accessor tests were added, so the tests module is now a sibling file
(host_state_tests.rs) included via #[cfg(test)] #[path = "..."] mod
tests. Parent file drops to 454 lines. No behavior change — it's a
mechanical split with the new test file carrying the existing 4 tests
plus the 2 new precedence tests.

All 256 astrid-capsule tests pass. cargo fmt --all --check clean.
